all bucket components are now functional components
